Add verbose debug logging for testing.
This adds the use of the OR_DEBUG environment variable to turn on
verbose logging.  Use with:

```
OR_DEBUG=true node index.js
```

Also turns this debugging on by default in tests, to facilitate more
useful output in travis build logs.  If we'd rather not have the
tests be noisy locally, we could rever the change to package.json
and just add the OR_DEBUG environment variable to travis config.